fix(components/Survey.js): useState instead of redux, changed the folder structure
