+++ src/Components/Survey/Buttons/ConfirmSurvey.js
@@ -1,6 +1,6 @@
 import { Checkbox, FormControlLabel } from "@material-ui/core";
 import React from "react";
-import i18n from '../../i18n';
+import i18n from '../../../i18n';
 
 export const ConfirmSurvey = ({ isConfirmed, handleConfirm, t }) =>
   <FormControlLabel
@@ -16,3 +16,5 @@ export const ConfirmSurvey = ({ isConfirmed, handleConfirm, t }) =>
     label={t('confirm survey')}
   />
 ;
+
+export default ConfirmSurvey;
