+++ src/Components/Survey/Buttons/RateSlider.js
