+++ src/Components/Survey/Buttons/TemperatureSlider.js
