+++ src/Components/Survey/Buttons/index.js
@@ -1,0 +1,15 @@
+import CheckBox from './CheckBox';
+import ConfirmSurvey from './ConfirmSurvey';
+import RateSlider from './RateSlider';
+import SubmitButton from './SubmitButton';
+import TemperatureSlider from './TemperatureSlider';
+import UnitRadioGroup from './UnitRadioGroup';
+
+export {
+    CheckBox,
+    ConfirmSurvey,
+    RateSlider,
+    SubmitButton,
+    TemperatureSlider,
+    UnitRadioGroup
+};
