+++ src/store/actions/actionTypes.js
@@ -1,7 +1,2 @@
 export const CHANGE_LANG = 'CHANGE_LANG';
 export const CHANGE_SCREEN = 'CHANGE_SCREEN';
-export const LOADING = 'LOADING';
-export const SUBMIT_FORM_SUCCESS = 'SUBMIT_FORM_SUCCESS';
-export const SUBMIT_FORM_FAILURE = 'SUBMIT_FORM_FAILURE';
-export const UNAUTHORIZED = 'UNAUTHORIZED';
-export const RESET_FORM_SUCCESS = 'RESET_FORM_SUCCESS';
