+++ src/store/reducers/index.js
@@ -3,10 +3,6 @@ import * as actionTypes from '../actions/actionTypes';
 const initailStata = {
     lan: 'English',
     curScreen: 'Home',
-    loading: false,
-    hasFetchingError: false,
-    formSuccess: false,
-    unhothorized: false,
 };
 
 const reducer = (state = initailStata, action) => {
@@ -21,51 +17,6 @@ const reducer = (state = initailStata, action) => {
             return {
                 ...state,
                 curScreen: action.screen
-            }
-        }
-        case actionTypes.LOADING: {
-            return {
-                ...state,
-                loading: true,
-                hasFetchingError: false,
-                formSuccess: false,
-                unhothorized: false,
-            }
-        }
-        case actionTypes.SUBMIT_FORM_SUCCESS: {
-            return {
-                ...state,
-                loading: false,
-                hasFetchingError: false,
-                formSuccess: true,
-                unhothorized: false,
-            }
-        }
-        case actionTypes.SUBMIT_FORM_FAILURE: {
-            return {
-                ...state,
-                loading: false,
-                hasFetchingError: true,
-                formSuccess: false,
-                unhothorized: false,
-            }
-        }
-        case actionTypes.RESET_FORM_SUCCESS: {
-            return {
-                ...state,
-                loading: false,
-                hasFetchingError: false,
-                formSuccess: false,
-                unhothorized: false,
-            }
-        }
-        case actionTypes.UNAUTHORIZED: {
-            return {
-                ...state,
-                loading: false,
-                hasFetchingError: false,
-                formSuccess: false,
-                unhothorized: true,
             }
         }
         default: {
